#40 refactor - moved cards into distinct folder.  social compoments moving via next commit.
