+++ components/Sidebar-left.js
@@ -13,8 +13,8 @@
 import { useState } from "react"
 import Link from "next/link"
 import { useLayout } from "./LayoutProvider"
-import ArtistCard from "/components/card_artist"
-import ListingCard from "/components/card_listing"
+import ArtistCard from "@/components/cards/card_artist"
+import ListingCard from "@/components/cards/card_listing"
 import Image from "next/image"
 import { useRouter } from "next/router"
 import { getRandomStockPhotoByCategory } from "@/utils/stockPhotos"
